Add files via upload
Add file temperature_dependent_materials with the classes used to reorganize the materials properties. Work to be completed. Need to complete the kapton class and go on.

Ad folder test_props with the tests performed until now to check the work. 
The __init__.py files are necessary to make the tests callable, otherwise imports error related to module temperature_dependent_materials.py arise.