feat: New method to get tsc in nodal or Gauss points.
New method to evaluate the current sharing temperature in nodal or in Gauss points according to the value of the flag nodal.
This is necessary to allow user to decide if current sharing temperature should be avaluated at each time steps or only at the inizializaion and at the end of the simulation trough the new flag TCS_EVALUATION to reduce computational time.

This is a very rough solution and should be improved ad soon as possible.