feat: new method eval_tcs.
New method eval_tsc that actually eavaluates the current sharing temperature in the selected location (nodal or Gauss points).

This method is a consequence of the possibility of the user to decide if the current sharing temperature should be evaluated at each time steps or only at the beginning and at the end of the simulation to reduce computational time.
This is a rough solution and should be impoved as soon as possible.