feat: new flag TCS_EVALUATION!
New flag that allows the user to decide if the current sharing temperature should be evaluated at each time steps or if it should be evaluated only at the beginning or at the end of the simulation.

This is a very rough solution and should be impoved as soon as possible.