feat!: new methods in class CylindricalHelix.
New private methods __eval_radius and __eval_alpha to evaluate the radius of the cylinder and the angle in the xy plane starting from the coordinates of the barycenter.
New method helix_parametrization to evaluate x, y and z spatial coordinates of the helix given the angluar discretization (tau).

Some updates in the docstring.