chore: update evaluation of quantities Q1 and Q2 in method eval_gauss_point of class Component
Keep into account of Joule power due to electric resistance along the strands and electric contuctance across the strands in the evaluation of arrays Q1 and Q2 used to build the source therms.

The contribution of Joule power due to electric resistance along the strands should be considered in function step: ask professor how to do this.