fix: convert conductor_definition input values in method __init__ of class Conductor
For sheet CONDUCTOR_input
Temporary solution to mangage input file loading, strange behavior: 1 are converted to True but 0 not converted to False. Works with no conversion only if the first value (XLENGHT) is set to 1.

For sheet CONDUTCTOR_operation
Temporary solution to mangage input file loading, strange behavior: 1 are converted to True but 0 not converted to False.