fix: RuntimeError in method solve_current_divider of class StrandMixedComponent
RuntimeError: Failed to converge after 1000 iterations, value is [].

This error is fixed deleting all the lines of code that are related to the evaluation of the electrical resistivity in nodal points in method get_electric_resistance of class StrandMixedComponent. Indeed, this properties is used to evaluate the electric resistance along current carriers and it is a properties associated to Gauss poins only.