chore: update private method __compute_node_distance in class Conductor
In this case attribute node_distance does not account for the twist pich of the strand, i.e. the strand is straight. Apply the costheta correction to account for the real distance between nodal points. This correction has an impact only in the evaluation of the quantity used in the electromagnetic module, as for instance electric resistance, electric conductance and inductance.
This is not necessary if there are more than one strand since in this case the helicoidal coordinates are used.