chore: update sheet STR_MIX
Added the possibility to explicitly define mixed strand (stabilizer + supercondutor) and segregated stabilizer (as the number of strand and the corresponding diameter for a specific type).

Stabilizer non stabilizer ratio can be of two kinds:
    1. accounts for segregated stabilizer;
    2. does not account for segredated stabilizer
This is used to correctly evaluate the superconduciting cross section and
therefore the total stabilizer cross section used in the electric module.

The scaling parameter for the critical current density correlation c0 can be according to two different definitions (flag C0_MODE):
    1. ingegneristic, i.e., the value is normalized with respect to the total perpendicular cross section of superconducting strand (stabilizer + superconductor);
    2. physical, i.e., the value is normalized with respect to the total perpendicular cross section of supecronductor only