fix: ValueError in current_sharing_temperature_nb3sn
ValueError: f(a) and f(b) must have different signs.

If the current is everywere equal to 0.0 A, the current sharing temperature is
equal to the critical temperature at the given magnetic field and strain by definition. Else if the current is not everywere equal to 0.0 A, set the current sharing temperature to the critical current at the given magnetic field and strain where the current is equal to 0.0 A and use bisection to find the current sharing temperature in all the other spatial coordinates.