feat: new function filter_component
Function that returns a new collection of tuple (index,object) with all the objects except the one in item; index are the locations of the objects in iterable. The returned collection (tuple) preserves the order of the items in iterable.

This function is used to remove if statements like if kk != jj in function step.