feat: new function array_initialization
Wrapper of function np.zeros that initializes array of given dimension according to the time step number.
N.B. the application of the theta method should be completely rivisited in the whole code.

This function is use to carry out refactoring of function step.
Code profiling is not perfomed at this stage.

modified: utility_functions/transient_solution_functions.py