refactor: replace [0] & [1] with namedtuple fields in function step
If conductor.cond_num_step = 1 SVEC and ELSLOD are namedtuple with fields previous and present that replace the old dictionary key 0 and 1. This should improve readability and maitainability of function step.

The value stored in previous and present fileds are ndarray vectors used to apply the theta method, which actually should be completely revised.

modified:   utility_functions/transient_solution_functions.py