chore: update function __build_fluid_eq_idx
The function now returns a dictionary of namedtuple with the equation index for all the FluidComponent objects. In this way it can be called only once inside funcion step since the indexes are constant. Actually, this property can be exploited to call this function only once in the code before calling function step.

This function is used to carry out refactoring of function step.
Code profiling is not perfomed at this stage.

modified:   utility_functions/transient_solution_functions.py