chore: update function __build_fluid_eq_idx
Function renamed to __build_equation_idx.
The function evaluates alos the index of the temperature equation of SolidComponent and stores them in the dictionary as integer value.

Docstring is updated accordingly.

This function is used to carry out refactoring of function step.
Code profiling is not perfomed at this stage.

modified:   utility_functions/transient_solution_functions.py