chore: new python module step_matrix_construction.py
Python module with functions used to initialize build and update the matrices used in function step to solve the thermal hydraulic problem.
This new module is defined downstream of prof. Alessandro Savino's suggestion in order to improve the readability and maintainability of the code.

new file:   utility_functions/step_matrix_construction.py