feat: new function build_svec
Function that builds the source vector (SVEC) elements at the Gauss point due to heat generation in strand and or jacket component objects and to thermal contact beween jacket compoments belonging to different conductors (qsource). For strand component objects the latter contribution is always zero.

N.B. This function is a merge of the if statement if isinstance(scomp,StrandComponent) is true do not account for qsource else, account for qsource. Since, as mentioned qsourse = 0 for StrandComponent, the check chan be avoided. This should improve readability and maintainability of the function.

This function is used to carry out refactoring of function step.
Code profiling is not perfomed at this stage.

modified:   utility_functions/step_matrix_construction.py