chore: add comment in function build_transport_coefficients
Added comment that specifies the rationale behind the loop on all the interfaces between fluid components.

The loop is done on all the type of interfaces between fluids since in the evaluation of K', K'' and K''' the relevant parameter is the open cross section per unit length (open contact perimeter): K', K'' and K''' are not 0 only if the open contact perimeter is not 0. In this way, a check on the interface kind between fluids is avoided for each interface and for each time step.

modified:   utility_functions/step_matrix_construction.py