feat: new function eval_system_matrix
Function that evaluates the system matrix using the values of the matrix MASMAT, FLXMAT, DIFMAT and SORMAT, according to the selected method for time integration.

This function is used to carry out refactoring of function step.
Code profiling is not perfomed at this stage.

modified:   utility_functions/step_matrix_construction.py