feat: new method build_th_bc_index in class FluidComponent
Method that builds and initialize the data structure to store the values of the index used to assign the boundary conditions at inlet and outlet for the thermal hydraulic problem. The data structure is a namedtuple with three keys (velocity pressure and temperature), each key has a namedtuple with two keys (forward and backward) to properly assing the boundary conditions in case of coolants in counter current flow.

This method is used to carry out refactoring of function step.
Code profiling is not perfomed at this stage.

modified:   fluid_component.py