refactor: fun build_equation_index -> method of Conductor
Function build_equation_index in module utility_functions/step_matrix_construction.py becomes a private method of class Conductor.

"Private method that evaluates the index of the velocity, pressure and temperature equation of the FluidComponent objects, collecting them in a dictionary of NamedTuple, together with the index of the temperature equation of the SolidComponent objects stored as integer in the same dictionary."

This is used to carry out refactoring of function step.
Code profiling is not perfomed at this stage.

modified:   utility_functions/step_matrix_construction.py
modified:   conductor.py