chore: call __build_equation_idx in Conductor __init__
Call method __build_equation_idx to build attribute equation_index;
self.equation_index -> dict: collection of NamedTuple with the index of velocity, pressure and temperaure equation for FluidComponent objects and of integer for the index of the temperature equation of SolidComponent. This is used in funcion step to solve the thermal hydraulic problem.

modified:   conductor.py