chore: call build_th_bc_index in Conductor.initialization
Call method build_th_bc_index of class FluidComponent to build namedtuples attributes inl_idx and inl_idx with the index used to assign the inlet and outlet BC.

This is used to carry out refactoring of function step.
Code profiling is not perfomed at this stage.

modified:   conductor.py