feat: new function impose_pressure_drop
Function that imposes a pressure drop as boundary conditions for the thermal hydraulic problem. Imposed quantities are:
    * inlet pressure
    * inlet temperature (or outlet temperature if back flow)
    * outlet pressure
The function suitably accounts for forward (left to right)/backward (right to left) flow as well as for back flow (part of the fluid moves towards the inlet due to a pressure build up).

This function is used to carry out refactoring of function step.
Code profiling is not perfomed at this stage.

modified:   utility_functions/th_boundary_conditions.py