feat: new function  save_ndarray
Function that saves selected ndarrays to perform test according to the TDD approach. The path to the directory where to save the ndarrays is hardcoded in the function. This is not optimum but it is not trivial to do it in a different way.

modified:   utility_functions/transient_solution_functions.py