chore: call function save_ndarray in function step
Called function save_ndarray in order to save ndarrays MASMAT, FLXMAT, DIFMAT, SORMAT, SYSMAT, SYSVAR, SYSLOD in function step before the application of BC. This can be useful to create and make tests according to TDD approach.

This lines of code are commented by default since this is not the standard behaviour of the functon step; in order to save the ndarrays user should decomment this lines.

modified:   utility_functions/transient_solution_functions.py