feat: new private function
Function that evaluates the S matrix (SMAT) therms due to solid component interfaces at the Gauss point (SOURCE JACOBIAN) by rows (horizontally) corresponding to the equations of comp_1 for the columns of comp_1 including the contributions given by comp_2.

This function is used to carry out refactoring of function step.
Code profiling is not perfomed at this stage.

Introduced to fix a mistake in the evaluation of the terms of SMAT matrix.

modified:   utility_functions/step_matrix_construction.py