chore: add flexibility in fixed_refined_spatial_discretization
Use new input variable DXINCRE_LEFT and DXINCRE_RIGHT from file conductor_grid.xlsx to improve the flexibility of the refined mesh. These variable are used to deal with the coarsening of the mesh to the left and to the right of the refined region respectively.

modified:   utility_functions/initialization_functions.py