feat: new function check_refined_grid_quality
Function that checks the quality of the refined grid. The check is performed comparing the spatial discretization pitch used for the uniform region (dx) to the left or to the right of the refined region, with respect to the last spatial discretization pitch used for coarsening (dx_coarse). If dx < dx_coarse the grid quality is not good and an error is raised.

modified:   utility_functions/initialization_functions.py