refactor: function get_time_step returns a time step
Previously the function updated inplace the value of the attribute conductor.time_step.
For clarity and for the sake of making the code explicit, the function now returns a float (the time step value) that should be assigned to the attribute conductor.time_step.

modified:   simulation.py
modified:   utility_functions/transient_solution_functions.py