feat!: new input file template_winding_pack_coupling.xlsx
New input file that allows user to specifiy the coupling between conductors that constitute a winding pack. The following sheets are available at the time being with the following meaning/usage:

    * contact_perimeter_flag: [-] contact perimeter flags between conductors: 1 = use value from sheet contact_perimeter; -1 = read from user defined auxiliary input file;
    * contact_perimeter: [m] wetted/contact perimeter if contact_perimeter_flag = 1
    * HTC_mode: [-] flag to define the model to compute the HTC: 1 = evaluated in the code accounting for jacket tickness and thermal conductivity; -1 = read from sheet contatc_HTC
    * contact_HTC: [W/m2K] Value for the heat transfer coefficient, if flag HTC_mode = -1
    * HTC_multiplier % [] Value for the multiplier for the heat transfer coefficient, when needed
    * thermal_contact_resistance: [m2K/W] thermal contact resistance between soldi component, should be >= 0; used if flag HTC_choice is 1
    * interf_thickness_mode: [-] Flag to select how to define the thickness of the interface between conductors in the winding pack: 1 = read from sheet interf_thick; -1 = read from user defined auxiliary input file
    * interf_thickness: [m] Thickness of the interface between conductor components if flag interf_thickness_mode = 1. The matrix may not be symmetric and the lower triangular region should be filled.

All the above sheets have a matrix form and can be considered symmetric so that only the upper triangular part should be filled with the exception of sheet interf_thickness that could not be symmetric since the thickness of the interface may be different from one conductor to the other.

BREAKING CHANGE: new input file that defines the windign pack (multi conductor model).

new file:   input_files/input_file_template/template_winding_pack_coupling.xlsx