chore: update funciton save_input_files
Add prefix meta (for metadata) to each file name to distinguish the
saved input files with respect to the initial set of input files used
to run the simulation. This allows to overcome the fact that two .xlsx
files with the same name could not be opened and could be useful for
file comparison.

modified:   simulation.py