chore: make method get_total_joule_power_electric_conductance public
Change status of method get_total_joule_power_electric_conductance from private to public in order to make it available outside of the class Conductor. This is necessary to call the method inside funcion electric_transient_solution.

This is something to be better addressed in a refactoring session. Probably the whole structure of the electric module should be reorganized to keep this and other methods private.

Docstring is updated accordingly.

Class Conductor
modified:   conductor.py