chore: update method electric_method
Call methods self.electric_solution_reorganization and self.get_total_joule_power_electric_conductance in the branch where the static solver for the electric method is used.
This should allow to correctily compute the Joule power and not mess up with the same evaluation in the transient solution.

Class: Conductor
modified:   conductor.py