chore: update method initialize_electric_quantities
Added the initialization of matrices total_linear_power_el_cond and linear_power_el_resistance used in the final step of the assembly of the source term array according to the numerical scheme for time integration (BE, CN or AM4).

Dostring is updated accordingly.

Class: SolidComponent.
modified:   solid_component.py