feat: new method get_joule_power_along_steady
Method that evaluate the contribution to the total power in the element of Joule power (in W/m) due to the electic resistances along the SolidComponent objects.
This method should be called in the electric method, when the transient solution is used and for the solid component that actually carries a current.
It works when the steady state solution for the electric module is computed.

Class: SolidComponent
modified:   solid_component.py