feat: new method get_joule_power_across_steady
Method that evaluates the contribution to the total power in the nodes of Joule power (in W/m) due to the electic conductance across the SolidComponent objects.
This method should be called in the electric method, when the transient solution is used.
It works when the steady state solution for the electric module is computed.

Class: SolidComponent
modified:   solid_component.py