feat: new private method __initialize_heat_source_nodal_pt_th
Private method that initializes heat source arrays in nodal points for strand and jacket objects keeping into account of the contribution strictly related to the thermal hydraulic model. The Joule power due to the electric module is evaluated in a different method. This method is called only once at conductor initialization and it is introduced in order to limit the number of times the check on time or number of step is carried out to select the suitable operation.

Class: Conductor
modified:   conductor.py