feat: new method eval_integral_joule_power
Method that evaluates the numerator of the expression used to evaluate the integral value of the Joule power. For the Joule power due to the electric resistance along the current carriers two different but equivalent approaches are used:
    * mode1 -> P_Joule = R * I^2 (R electric resistance in Ohm, I electric current)
    * mode2 -> P_Joule = \Delta_Phi * I
with
    * R electric resistance in Ohm
    * I electric current in A
    * \Delta_Ph electric voltage potential difference along current carriers in V
For the Joule power due to the electric conductances between current carriers it is exploited the power computed in method get_total_joule_power_electric_conductance
Regardless of the kind of Joule power, the integration can be performed as follows:
    P_Joule = 1/Delta_t_TH * int_0^Delta_t_TH (dt P_{Joule,i})
The above integral could be approximated as
    P_Joule ~= 1/Delta_t_TH * sum_1^N_em dt_em P_{Joule,i}
Where:
    * P_Joule: integral value of the Joule power in W
    * P_{Joule,i}: value of the Joule power computed with the electric
        soltution at the i-th electric time step in W
    * Delta_t_TH: value of the thermal hydraulic time step in s (always >=
electric time step)
    * dt_em: electric time step in s
    * N_em: number of electric time step required to cover a full thermal
    hydraulic time step.

This method computes the numerator of the above equation, i.e.
    sum_1^N_em dt_em P_{Joule,i}
The final quantities is computed in other methods (get_joule_power_along, get_joule_power_across).

Class: Conductor
modified:   conductor.py