chore: update method get_joule_power_along
Method self.get_joule_power_along is used to evalutate the integral (in time) value of the Joule power due to the electric resistances along current carriers.
The procedure is started in the transient electric loop with the evaluation of the numerator (the sum of all the products P_{Joule,i} * dt_em that could be computed at each electric time steps) and it is finalized in this method by dividing the sum by the product of the thermal hydraulic time step and of the element length to get a linear power density (W/m).

The method should be called whithin the electric method, when the transient solution is used and only for current carriers (removed check on attribute self.name).

Class: SolidComponent
modified:   solid_component.py