chore: update method __build_heat_source_nodal_pt
Removed call to methods obj.get_joule_power_along and obj.get_joule_power_across for jacket components since at the time being they are not supposed to carry a current.
Those methods should now be called only if the component carries a current since the check on attribute obj.name was remove.
If they are used on not current carrier components, being all the variables set to zero, the operation will result in an array that is still full of zeros but it slowes down the computation.

Class: Conductor
modified:   conductor.py