doc: update documentation in file template_conductor_1_operation
Added comment related to the necessity of specify the value of TQBEG in the input file also if the heating is read from intput file since it will somehow influence the evaluation of the heat transfer coefficient between fluid and solid component if there is an interface and if it is evaluated by the software. A better solution should be found but for the time being it is a good quick fix.

modified:   input_files/input_file_template/template_conductor_1_operation.xlsx