chore: add check in method eval_transp_coeff
Added check on flag IQFUN for the evaluation of htc_transient.
The check allows to compute htc_transient only if there is an heating. At the time being this correction does not account for the different ways to define an external heating, that is why user should specify the value of TQBEG also if IQFUN = -1 (read heat from external file).

Class: Conductor
modified:   conductor.py