feat: implemented Class stack, only for some cases
The class stack is implemented following the same structure of the strand mix class.
The only limit of the new class is related to the evaluations of the self and the mutual inductance, since the old method takes into account the radius of the strand.
In this method you must impose a self and mutual inductance.
For the future a general way to evaluate these parameters must be implemented.
That's the reason why also the conductor.py file has been changed.
The new class is able to represent 4 different materials inside the stack.
Homogenizes non-superconducting materials into a single stab material and performs the same electrical calculations present in the strand mix class.
	modified:   conductor.py
	modified:   stack_component.py