chore: remove mod1 to compute Joule power along
During the fix of the evaluation of the Joule power along the current carrier, two different approaches were used:

    * mod1 -> P_joule = R * I^2
    * mod2 -> P_joule = Delta_Phi * I

The approaches should be equivalent if the inducance is negligible, while mod2 is more general and conservative if the inductance is no longer negligible. After discussion with prof. Zach Hartwig and Dr. Nicolò Riva, we agreed to remove mod1 and keep mod2 for this evaluation.

List of changes:

    * removed key "integral_power_el_res_mod1" from dictionary obj.dict_Gauss_pt of current carrier objects.
    * renamed key "integral_power_el_res_mod2" into "integral_power_el_res" in dictionary obj.dict_Gauss_pt of current carrier objects.
    * removed equivalence checks
    * updated related comments
    * updated docstring

LList of modified methods in class SolidComponents:

  * initialize_electric_quantities
  * set_power_array_to_zeros
  * get_joule_power_along
  * get_joule_power_along_steady

List of modified methods in class Conductor:

  * __electric_method_steady
  * __build_heat_source_gauss_pt
  * eval_integral_joule_power

Class: SolidComponents, Conductor
modified:   conductor.py
modified:   solid_component.py