fix: update temporary solution input loading in Conductor
Updated the instruction that manages the odd behaviour while loading input
files (value set to 1 are converted in True) in method __init__.
To avoid that other values are converted unexpectedly, the explicit check
for value == True is added in this commit.

For instance this allows to avoid the conversion of variable INDUCTANCE_MODE = 2
into INDUCTANCE_MODE = 1 (because 2 is equivalent to true but is different from it).

Class: Conductor
modified:   conductor.py