feat!: new variables in file conductor_definitions.xlsx
The new variables are:
    * MUTUAL_INDUCTANCE -> Constant value for the mutual inductance used if flag INDUCTANCE_MODE is set to 0.
    * SELF_INDUCTANCE -> Constant value used if SELF_INDUCTANCE_MODE is set to 0.

New flag values:
    * INDUCTANCE_MODE = 0 -> constant value from variable MUTUAL_INDUCTANCE
    * SELF_INDUCTANCE_MODE = 0 -> constant value from variable SELF_INDUCTANCE

This commit was already available in the pull request from Alessandro Mortara but gives conflict that was not trivial to resolve since it was not possible to visualize them with the text editor or with GitHub Desktop. So I add the same changes he proposed in this commit.

BREAKING CHANGES
    * New variables MUTUAL_INDUCTANCE, SELF_INDUCTANCE
    * SELF_INDUCTANCE_MODE flag is used whatever is the flag INDUCTANCE_MODE
    * Default value for flag SELF_INDUCTANCE_MODE is changed to 1

modified:
input_files/input_file_template/template_conductor_definition.xlsx