feat!: new module adaptive_mesh_functions.py
Added new python module where all the functions used to deal with the adaptive mesh are collected.

N.B. The adaptive mesh that is going to be implemented does not belong to any of the rigorous adaptive mesh schemes that can be found in literature (h-adaptivity, p-adaptivity, r-adaptivity). It is
an ad hoc algorithm to generate an adaptive mesh which works for the kind of problems that can be addressed with the OpenSc2. For further details about the algorithm, please write an e-mail to laura.savoldi@polito.it.

new file:   utility_functions/adaptive_mesh_functions.py