feat: new function update_mesh
Function that identifies regions of the mesh that need refinement and regions of the mesh that need coarsening, performing the corresponding adjustment of the mesh.
Regions requiring coarsening or refinement are identified by comparing the actual mesh density wrt the "ideal" mesh density computed with the eval_gaussian_mesh_density function.
For each element of the mesh:
    * coarsening is performed by calling the coarse_mesh function;
    * refinement is performed by calling the refine_mesh function;
    * regions that do not require coarsening/refinement are treated with the set_node function.

modified:   utility_functions/adaptive_mesh_functions.py