feat: new function set_node
Function that adds a new node in the new mesh according to the info available
in the old mesh. The added node corresponds to the upper boundary of the j-th
element of the mesh (at index j+1). The node is characterized as soft or hard
according to the information stored in key "hard_node_value" of dictionary
grid_feat.

N.B. Input argument grid_input is not used but it is required in order to have
the function set_node, refine_mesh and coarse_mesh with the same signature for
future refactoring.

modified:   utility_functions/adaptive_mesh_functions.py