feat: new function coarse_mesh
Function that locally coarsen the mesh removing nodes only if they are soft. If
nodes are hard, they are not removed as they correspond to the nodes of the
initial mesh.
The set_node function is also called to correctly deal with the end node of the
element.

modified:   utility_functions/adaptive_mesh_functions.py