chore: new attr tau in instances of Conductor components
Added new attribute tau (angular discretization) to instances of Conductor
component objects used to correctly update the coordinates of barycenter if the
mesh is adaptive.
Updated updated functions:
    * helicoidal_coordinates
    * build_coordinates_of_barycenter

Heavy refactoring is needed in this part of the OpenSc2 that deals with 3D
coordinate generation.

modified:   utility_functions/initialization_functions.py