feat: new function adaptive_mesh
New function that manages the mesh adaptivity according to an algorithm based
on quench front detection. When the temperature of an instance of class
StackComponent or StrandMixedComponent crosses the current sharing temperature,
a quench front is found. The regions in the nearby a quench front are typically
characterized by strong gradients and needs mesh refinement. Thus, in those
regions the mesh is locally refined adding (soft) nodes. Since the quench front
propagates, the added soft nodes may also be removed when they are no longer
needed. This process is called mesh coarsening. For more info on the
algorithm contact laura.savoldi@polito.it.

modified:   utility_functions/adaptive_mesh_functions.py