chore: call adaptive_mesh in method conductor_solution
Called method adaptive_mesh in method conductor_solution of class Simulation
if the mesh is adaptive. The function updates the mesh and all relevant mesh
parameters calling function adaptive_mesh. The function update also parameters
and quantities that depend on the mesh and/or on the number of nodes/elements
of the mesh.

Class: Simulation
modified:   simulation.py