chore: eval zcoord_gauss only at t = 0.0 s
Updated private method self.__update_grid_features of class Conductor.

Array zcoord_gauss is evaluated by this functin only at the
instatiation/initialization phase. For the remaining of the simulation, this
method is called only if the mesh is adaptive. If it is the case, the updated
values of zcoord_gauss are computed in function update_mesh and temporarily
stored in array zcoord_gauss_new, which is used to perform interpolations. When
all interpolations are done, array zcoord_gauss is updated with values in
zcoord_gauss_new calling method self.update_cond_mesh_related_features in
function adapt_mesh.

Class: Conductor
modified:   conductor.py