feat: new method interp_tcs_on_new_mesh in Conductor
New method that interpolates the current sharing temperature on the new
mesh if flag TCS_EVALUATION is set to False in both nodal and Gauss points,
in order to evaluate material properties that depend on this parameter. This is
done to avoid a mismatch between the size of the temperature array and the size
of the current sharing temperature array. The minimum current sharing
temperature for the time being is set equal to the current sharing temperature
as already done in method get_tcs of class Strand.
If flag TCS_EVALUATION is set to True, the TCS is evaluated before the
calculation of the other material properties and no mismatch is found.
Attribute dictionaries dict_node_pt and dict_Gauss_pt of instances of class
StackComponent and StrandMixedComponent are updated inplace.

Class: Conductor
modified:   conductor.py