fix: ValueError in isobaric_specific_heat_nb3sn
ValueError: operands could not be broadcast together with shapes (115,) (101,)

This error is due to a not correct update of array T_curr_sharing_min on the
new mesh (if the mesh is adaptive). Calling method self.interp_tcs_on_new_mesh
in function adaptive mesh, array T_curr_sharing_min is correctly interpolated
on the new mesh and the bug is fixed.
This explains why the bug is detected in function isobaric_specific_heat_nb3sn and
it is fixed in functon adaptive_mesh.

modified:   utility_functions/adaptive_mesh_functions.py