fix: IndexError in build_svec
IndexError: index 101 is out of bounds for axis 0 with size 101

This bug is related to not interpolation on the new mesh of attribute
dictionary dict_qsource of class Simulation, which stores the heat exchanged
between conductors if simulation with multiconductor is performed.

This bug is fixed calling method interp_heat_btw_cond_on_new_mesh of class
Simulation in function adaptive_mesh. This is a workaround since the
multiconductor is still not available in OpenSc2.

modified:   utility_functions/adaptive_mesh_functions.py