feat: new method update_delta_voltage_along_on_new_mesh
New method that updates the voltage difference along on the new mesh. If the
simulation if purely thermal hydraulic, the voltage is zero in along the whole
conductor, so array delta_voltage_along is set to 0 accounting for the new
number of nodes in the mesh.

Polymorphism is exploited.

In class SolidComponen (parent class) the method raises a NotImplementedError
as the real behavior of the method is specified in classes StrandComponent and
JacketComponent.

In class StrandComponent the method checks that the electric method is not used,
i.e. the simulation is purely thermal hydraulic. If it is the case, it returns
the updated value of ndarray delta_voltage_along set to zeros and with the shape
consistent to the shape of ndarray zcoord_gauss.

In class JacketComponent the method simply returns the updated value of ndarray
delta_voltage_along set to zeros and with the shape consistent to the shape of
ndarray zcoord_gauss.

Class: SolidComponen, StrandComponent, JacketComponent
modified:   jacket_component.py
modified:   solid_component.py
modified:   strand_component.py