fix: ValueError in store_interp_spatial_distributions
ValueError: operands could not be broadcast together with shapes (100,) (114,)

The bug is due to the not correct update of the size of ndarray
delta_voltage_along in instances of class StrandComponent.

It is fixed calling the new method update_delta_voltage_along_on_new_mesh of
Class SolidComponent in functon adaptive_mesh. The method performs this update
according to the kind of object exploiting polymorphism.

modified:   utility_functions/adaptive_mesh_functions.py