fix: ValueError in interp_at_t_save
ValueError: operands could not be broadcast together with shapes (100,) (114,)

The bug is due to the not update on the new mesh of ndarray
linear_power_el_resistance for instances of class Jacket.
The bug is fixed adding a re-initialization to 0 with the correct shape in
method interp_joule_power_on_new_mesh of class Conductor for instances of class
JacketComponent.

Class: Conductor
modified:   conductor.py