fix: ValueError in _compute_density_and_mass_flow_rates
ValueError: No outputs were able to be calculated

This error is due to the fact that, if adaptive mesh is used, in method
update_cond_mesh_related_features of class Conductor the ndarray stored in
key SYSVAR of attribute dictionary dict_Step was set to zero for
"re-initialization" with the correct shape.
This has an effect on the evaluation of the known therm vector of the thermal
hydraulic problem that affects the solution leading to NaN values that triggers
the ValueError in CoolProp (uset to compute the density in method
_compute_density_and_mass_flow_rates of class FluidComponent).

To fix this bug the following actions were carried out:
    1. in method interp_solution_on_new_mesh of class Conductor, the
    interpolation of the whole solution is done on SYSVAR and not on
    SYSVAR_old, sice this key is used in funcion step to build the known
    therm vector. Array SYSVAR_old is created as an hard copy of SYSVAR in the
    same method.
    2. in method update_cond_mesh_related_features of Class Conductor, the
    statement that re-initializes to zero array SYSVAR is deleted. As a
    consequence, the method returns only 2 dictionaryes (grid_input and
    grid_features).
    3. To be consistent to the change of the number of output of method
    update_cond_mesh_related_features, in function adaptive_mesh is modified
    accordingly.

Docstrigns are modified accordingly where needed.

Class: Conductor
modified:   conductor.py
modified:   utility_functions/adaptive_mesh_functions.py